Fix combo-box not selecting the current data when the cell is edited
The concurrent run of the query made the setEditorData function to not
find the current value in the combo-box, because it was not actually any
data loaded. Using completeCache() the data is loaded before trying to set
the data in the editor.

To minimise possible performance problems, the model only loads the foreign
key field, instead of the whole related table.