Make Ctrl+T shortcut to work for "Open tab" in "Execute SQL"
This shortcut was set for the following actions, making it ambiguous:
"View > DB Toolbar" and "Open tab" in "Execute SQL". Additionally it was
used by QScintilla for transposing lines.

Shortcut for "View > DB Toolbar" has been removed, since it's not
considered a frequently used action and there are other toolbars without
shortcut. The shortcut for QScintilla has been changed to "Ctrl+Shift+Up".

The now working shortcut is added to the "Open tab" action's tool-tip for
documentation purposes.

See issue #1647