Fix tabs not showing in a clean installation
If there are no settings previously stored for the MainWindow/openTabs, the
default empty string was converted to a list of one empty string and
consequently all tabs were cleared and none added.

This fixes all possible cases where the list of open tabs is empty. It
means: do not make changes, as originally designed (since the setting for
all tabs closed doesn't make sense).

See issue #1692 (and #1675 for the enhancement).