Fix ambiguous shortcut for Ctrl+/
The shortcut was ambiguous because it was active in the button and in the
SQL editor itself. In the button was only documentation and was made
ambiguous when enabling the feature and fixing the shortcut in
6425fb177eba293ae000415392d088af96e27c91

Now its scope is reduced to widget like similar button shortcuts so it can
be unambiguous in the editor.

See issue #1614