Check number of columns in custom display format
Added a new callback for executeSQL following the model of sqlite3_exec.
Using this, the number of columns is got from the query execution. If it
is not one, the custom display format is rejected.

The regexp is also narrowed to accept only SQL functions applied to the
column name (it now matches the entire string).

See issue #573