Don't show misleading "determining row count" message in empty db
When no table is selected in the Browse Data tab, either because the
current database is empty or because no database is opened, we would
forever show the "determining row count" message even though nothing is
to be determined at this point. This commit makes sure to show a better
"0 - 0 of 0" message as we did before.

See issue #1654.

# Conflicts:
#	src/sqlitetablemodel.cpp