Use the same code for creating new DBs as for opening existing DBs
We didn't install all the collations and callbacks we have when creating
a new database. They were only installed when opening an existing
database.

To avoid issues like this the code for creating a new database is
stripped down to a bare minimum. After the database is created, we
immediately close it now and open it again through the standard open
code.