Reduce (and unify) margins in the main tabs and in Execute SQL widgets
The margins have been set to 3 for the layouts of all the main tabs. It has
been generally reduced from the default (9), except for the DB Structure
which was set to 0 and now has been increased to 3 for uniformity with the
other tabs.

The margins of interior widgets in Execute SQL tab has also been reduced to
3 from the default.

See issues #1762 and #620.