Avoid style-sheet inheritance in colour chooser boxes in Preferences
This avoid propagating the style-sheet to other widgets. For example, to
the tool-tip shown in the frame. But it could also fixed the problem
reported for Windows in issue #1493:
"Editing color preferences dialog background takes the last chosen color."
"Color dialog becomes obscured by the changing color patterns of each
selection."