Highlight all the SQLite keywords and only those
Some SQLite keywords were not highlighted and QScintilla included a lot
of words that were not keywords for SQLite. This redefines the keyword
list for highlighting to match the list that we already prepare for
completion.

The list has been updated with some missing cases from the current SQLite
keyword list: https://www.sqlite.org/lang_keywords.html