Use a custom extension function instead of the json_array function
For multi-column PKs in WITHOUT ROWID tables the first implementation
added a dependency on the JSON extension by using the json_array
function. To make sure DB4S works even without this extension, this
commit adds and makes use of a custom extension function which
essentially does the same thing as json_array.